Use the queryparam to set the selection drop down
The idea is that when user navigates back to this page, they shall see what they had selected in the drop down. The bestuurseenheidId property keeps track of that selection. Here, we wrap that id value inside this route's model object. Then in the template we pass model.id to the power-select component.